Make it so that astropy.wcs.setup_package can be imported and used without importing the rest of astropy
