+++ astropy/wcs/wcs.py
@@ -45,7 +45,7 @@ import numpy as np
 # LOCAL
 from astropy import log
 from astropy.io import fits
-from . import _docutil as __
+from . import docstrings as docutils
 try:
     from . import _wcs
 except ImportError:
@@ -1359,9 +1359,9 @@ reduce these to 2 dimensions using the naxis kwarg.
 
         InvalidTransformError
             Ill-conditioned coordinate transformation parameters.
-        """.format(__.TWO_OR_MORE_ARGS('naxis', 8),
-                   __.RA_DEC_ORDER(8),
-                   __.RETURNS('sky coordinates, in degrees', 8))
+        """.format(docutils.TWO_OR_MORE_ARGS('naxis', 8),
+                   docutils.RA_DEC_ORDER(8),
+                   docutils.RETURNS('sky coordinates, in degrees', 8))
 
     def wcs_pix2world(self, *args, **kwargs):
         if self.wcs is None:
@@ -1428,9 +1428,9 @@ reduce these to 2 dimensions using the naxis kwarg.
         `~astropy.wcs.Wcsprm.lattyp` and `~astropy.wcs.Wcsprm.lngtyp`
         members can be used to determine the order of the axes.
 
-        """.format(__.TWO_OR_MORE_ARGS('naxis', 8),
-                   __.RA_DEC_ORDER(8),
-                   __.RETURNS('world coordinates, in degrees', 8))
+        """.format(docutils.TWO_OR_MORE_ARGS('naxis', 8),
+                   docutils.RA_DEC_ORDER(8),
+                   docutils.RETURNS('world coordinates, in degrees', 8))
 
     def _all_world2pix(self, world, origin, tolerance, maxiter, adaptive,
                        detect_divergence, quiet):
@@ -2164,9 +2164,9 @@ reduce these to 2 dimensions using the naxis kwarg.
         requested accuracy.  After 6 iterations, the solution is
         diverging at least for one input point.
 
-        """.format(__.TWO_OR_MORE_ARGS('naxis', 8),
-                   __.RA_DEC_ORDER(8),
-                   __.RETURNS('pixel coordinates', 8))
+        """.format(docutils.TWO_OR_MORE_ARGS('naxis', 8),
+                   docutils.RA_DEC_ORDER(8),
+                   docutils.RETURNS('pixel coordinates', 8))
 
     def wcs_world2pix(self, *args, **kwargs):
         if self.wcs is None:
@@ -2227,9 +2227,9 @@ reduce these to 2 dimensions using the naxis kwarg.
 
         InvalidTransformError
             Ill-conditioned coordinate transformation parameters.
-        """.format(__.TWO_OR_MORE_ARGS('naxis', 8),
-                   __.RA_DEC_ORDER(8),
-                   __.RETURNS('pixel coordinates', 8))
+        """.format(docutils.TWO_OR_MORE_ARGS('naxis', 8),
+                   docutils.RA_DEC_ORDER(8),
+                   docutils.RETURNS('pixel coordinates', 8))
 
     def pix2foc(self, *args):
         return self._array_converter(self._pix2foc, None, *args)
@@ -2258,8 +2258,8 @@ reduce these to 2 dimensions using the naxis kwarg.
 
         ValueError
             Invalid coordinate transformation parameters.
-        """.format(__.TWO_OR_MORE_ARGS('2', 8),
-                   __.RETURNS('focal coordinates', 8))
+        """.format(docutils.TWO_OR_MORE_ARGS('2', 8),
+                   docutils.RETURNS('focal coordinates', 8))
 
     def p4_pix2foc(self, *args):
         return self._array_converter(self._p4_pix2foc, None, *args)
@@ -2287,8 +2287,8 @@ reduce these to 2 dimensions using the naxis kwarg.
 
         ValueError
             Invalid coordinate transformation parameters.
-        """.format(__.TWO_OR_MORE_ARGS('2', 8),
-                   __.RETURNS('focal coordinates', 8))
+        """.format(docutils.TWO_OR_MORE_ARGS('2', 8),
+                   docutils.RETURNS('focal coordinates', 8))
 
     def det2im(self, *args):
         return self._array_converter(self._det2im, None, *args)
@@ -2316,8 +2316,8 @@ reduce these to 2 dimensions using the naxis kwarg.
 
         ValueError
             Invalid coordinate transformation parameters.
-        """.format(__.TWO_OR_MORE_ARGS('2', 8),
-                   __.RETURNS('pixel coordinates', 8))
+        """.format(docutils.TWO_OR_MORE_ARGS('2', 8),
+                   docutils.RETURNS('pixel coordinates', 8))
 
     def sip_pix2foc(self, *args):
         if self.sip is None:
@@ -2357,8 +2357,8 @@ reduce these to 2 dimensions using the naxis kwarg.
 
         ValueError
             Invalid coordinate transformation parameters.
-        """.format(__.TWO_OR_MORE_ARGS('2', 8),
-                   __.RETURNS('focal coordinates', 8))
+        """.format(docutils.TWO_OR_MORE_ARGS('2', 8),
+                   docutils.RETURNS('focal coordinates', 8))
 
     def sip_foc2pix(self, *args):
         if self.sip is None:
@@ -2394,8 +2394,8 @@ reduce these to 2 dimensions using the naxis kwarg.
 
         ValueError
             Invalid coordinate transformation parameters.
-        """.format(__.TWO_OR_MORE_ARGS('2', 8),
-                   __.RETURNS('pixel coordinates', 8))
+        """.format(docutils.TWO_OR_MORE_ARGS('2', 8),
+                   docutils.RETURNS('pixel coordinates', 8))
 
     def to_fits(self, relax=False, key=None):
         """
